@@ -5,16 +5,18 @@
 
 #pragma once
 
-#include "QuantumComputation.hpp"
+#include "Definitions.hpp"
 #include "cliffordsynthesis/Configuration.hpp"
 #include "cliffordsynthesis/Results.hpp"
 #include "cliffordsynthesis/Tableau.hpp"
+#include "cliffordsynthesis/TargetMetric.hpp"
 #include "cliffordsynthesis/encoding/SATEncoder.hpp"
-#include "plog/Log.h"
+#include "ir/QuantumComputation.hpp"
 
 #include <cstddef>
 #include <limits>
 #include <memory>
+#include <plog/Log.h>
 #include <sstream>
 #include <utility>
 
@@ -37,7 +39,7 @@ public:
         initialCircuit(std::make_shared<qc::QuantumComputation>(qc)),
         results(qc, targetTableau) {}
   explicit CliffordSynthesizer(qc::QuantumComputation& qc,
-                               const bool              useDestabilizers = false)
+                               const bool useDestabilizers = false)
       : initialTableau(qc.getNqubits(), useDestabilizers),
         targetTableau(qc, 0, std::numeric_limits<std::size_t>::max(),
                       useDestabilizers),
@@ -69,16 +71,16 @@ public:
   }
 
 protected:
-  Tableau                                 initialTableau{};
-  Tableau                                 targetTableau{};
-  std::shared_ptr<qc::QuantumComputation> initialCircuit{};
+  Tableau initialTableau;
+  Tableau targetTableau;
+  std::shared_ptr<qc::QuantumComputation> initialCircuit;
 
   Configuration configuration{};
 
-  Results                                 results{};
-  std::shared_ptr<qc::QuantumComputation> resultCircuit{};
-  Tableau                                 resultTableau{};
-  std::size_t                             solverCalls{};
+  Results results;
+  std::shared_ptr<qc::QuantumComputation> resultCircuit;
+  Tableau resultTableau;
+  std::size_t solverCalls{};
 
   static bool requiresMultiGateEncoding(const TargetMetric metric) {
     return metric == TargetMetric::Depth;
@@ -86,8 +88,8 @@ protected:
 
   void determineInitialTimestepLimit(EncoderConfig& config);
   std::pair<std::size_t, std::size_t> determineUpperBound(EncoderConfig config);
-  void                                runMaxSAT(const EncoderConfig& config);
-  Results                             callSolver(const EncoderConfig& config);
+  void runMaxSAT(const EncoderConfig& config);
+  Results callSolver(const EncoderConfig& config);
 
   void minimizeGatesFixedDepth(EncoderConfig config);
 
@@ -99,7 +101,7 @@ protected:
   void twoQubitGateOptimalSynthesis(EncoderConfig config, std::size_t lower,
                                     std::size_t upper);
 
-  void minimizeTwoQubitGatesFixedGateCount(std::size_t   gateCount,
+  void minimizeTwoQubitGatesFixedGateCount(std::size_t gateCount,
                                            EncoderConfig config);
   void minimizeGatesFixedTwoQubitGateCount(EncoderConfig config);
 
@@ -155,7 +157,7 @@ protected:
                        const Configuration& config);
   static void updateResults(const Configuration& config,
                             const Results& newResults, Results& currentResults);
-  void        removeRedundantGates();
+  void removeRedundantGates();
 };
 
 } // namespace cs
