@@ -5,30 +5,21 @@
 
 #pragma once
 
-#include "CircuitOptimizer.hpp"
 #include "cliffordsynthesis/Tableau.hpp"
+#include "ir/QuantumComputation.hpp"
 #include "logicblocks/Logic.hpp"
 
-#include <nlohmann/json.hpp>
-#include <sstream>
+#include <cstddef>
+#include <limits>
+#include <nlohmann/json_fwd.hpp>
+#include <ostream>
 #include <string>
 
 namespace cs {
 class Results {
 public:
   Results() = default;
-  Results(qc::QuantumComputation& qc, const Tableau& tableau) {
-    // SWAP gates are not natively supported in the encoding, so we need to
-    // decompose them into sequences of three CNOTs.
-    qc::CircuitOptimizer::decomposeSWAP(qc, false);
-
-    setResultCircuit(qc);
-    setResultTableau(tableau);
-    setDepth(qc.getDepth());
-    setSingleQubitGates(qc.getNsingleQubitOps());
-    setTwoQubitGates(qc.getNindividualOps() - singleQubitGates);
-    setSolverResult(logicbase::Result::SAT);
-  }
+  Results(qc::QuantumComputation& qc, const Tableau& tableau);
 
   virtual ~Results() = default;
 
@@ -39,8 +30,8 @@ public:
   [[nodiscard]] std::size_t getSingleQubitGates() const {
     return singleQubitGates;
   }
-  [[nodiscard]] std::size_t       getDepth() const { return depth; }
-  [[nodiscard]] double            getRuntime() const { return runtime; }
+  [[nodiscard]] std::size_t getDepth() const { return depth; }
+  [[nodiscard]] double getRuntime() const { return runtime; }
   [[nodiscard]] logicbase::Result getSolverResult() const {
     return solverResult;
   }
@@ -56,16 +47,8 @@ public:
   void setSolverResult(const logicbase::Result r) { solverResult = r; }
   void setSolverCalls(const std::size_t c) { solverCalls = c; }
 
-  void setResultCircuit(qc::QuantumComputation& qc) {
-    std::stringstream ss;
-    qc.dumpOpenQASM3(ss);
-    resultCircuit = ss.str();
-  }
-  void setResultTableau(const Tableau& tableau) {
-    std::stringstream ss;
-    ss << tableau;
-    resultTableau = ss.str();
-  }
+  void setResultCircuit(qc::QuantumComputation& qc);
+  void setResultTableau(const Tableau& tableau);
 
   [[nodiscard]] bool sat() const {
     return getSolverResult() == logicbase::Result::SAT;
@@ -74,33 +57,20 @@ public:
     return getSolverResult() == logicbase::Result::UNSAT;
   }
 
-  [[nodiscard]] virtual nlohmann::json json() const {
-    nlohmann::json resultJSON{};
-    resultJSON["solver_result"]      = toString(solverResult);
-    resultJSON["single_qubit_gates"] = singleQubitGates;
-    resultJSON["two_qubit_gates"]    = twoQubitGates;
-    resultJSON["depth"]              = depth;
-    resultJSON["runtime"]            = runtime;
-    resultJSON["solver_calls"]       = solverCalls;
+  [[nodiscard]] virtual nlohmann::basic_json<> json() const;
 
-    return resultJSON;
-  }
-
-  friend std::ostream& operator<<(std::ostream& os, const Results& config) {
-    os << config.json().dump(2);
-    return os;
-  }
+  friend std::ostream& operator<<(std::ostream& os, const Results& config);
 
 protected:
-  logicbase::Result solverResult     = logicbase::Result::NDEF;
-  std::size_t       singleQubitGates = std::numeric_limits<std::size_t>::max();
-  std::size_t       twoQubitGates    = std::numeric_limits<std::size_t>::max();
-  std::size_t       depth            = std::numeric_limits<std::size_t>::max();
-  double            runtime          = 0.0;
-  std::size_t       solverCalls      = 0U;
+  logicbase::Result solverResult = logicbase::Result::NDEF;
+  std::size_t singleQubitGates = std::numeric_limits<std::size_t>::max();
+  std::size_t twoQubitGates = std::numeric_limits<std::size_t>::max();
+  std::size_t depth = std::numeric_limits<std::size_t>::max();
+  double runtime = 0.0;
+  std::size_t solverCalls = 0U;
 
-  std::string resultTableau{};
-  std::string resultCircuit{};
+  std::string resultTableau;
+  std::string resultCircuit;
 };
 
 } // namespace cs
