@@ -7,10 +7,13 @@
 
 #include "cliffordsynthesis/Results.hpp"
 #include "cliffordsynthesis/Tableau.hpp"
+#include "ir/operations/OpType.hpp"
+#include "logicblocks/Logic.hpp"
 #include "logicblocks/LogicBlock.hpp"
 
 #include <cstddef>
 #include <memory>
+#include <utility>
 
 namespace cs::encoding {
 
@@ -18,18 +21,18 @@ class TableauEncoder {
 public:
   TableauEncoder() = default;
   TableauEncoder(const std::size_t nQubits, const std::size_t tableauSize,
-                 const std::size_t                      timestepLimit,
+                 const std::size_t timestepLimit,
                  std::shared_ptr<logicbase::LogicBlock> logicBlock)
       : N(nQubits), S(tableauSize), T(timestepLimit),
         lb(std::move(logicBlock)) {}
 
   struct Variables {
     // variables for the X parts of the tableaus
-    logicbase::LogicMatrix x{};
+    logicbase::LogicMatrix x;
     // variables for the Z parts of the tableaus
-    logicbase::LogicMatrix z{};
+    logicbase::LogicMatrix z;
     // variables for the phase parts of the tableaus
-    logicbase::LogicVector r{};
+    logicbase::LogicVector r;
 
     // update rules for single-qubit gates
     [[nodiscard]] logicbase::LogicTerm
@@ -75,6 +78,6 @@ protected:
   Variables vars{};
 
   // the logic block to use
-  std::shared_ptr<logicbase::LogicBlock> lb{};
+  std::shared_ptr<logicbase::LogicBlock> lb;
 };
 } // namespace cs::encoding
