@@ -8,7 +8,7 @@
 #include "Definitions.hpp"
 #include "NeutralAtomArchitecture.hpp"
 #include "NeutralAtomDefinitions.hpp"
-#include "operations/Operation.hpp"
+#include "ir/operations/Operation.hpp"
 
 #include <cstdint>
 #include <map>
@@ -18,43 +18,43 @@
 
 namespace na {
 class AnimationAtoms {
-  using axesId   = std::uint32_t;
+  using axesId = std::uint32_t;
   using marginId = std::uint32_t;
 
 protected:
-  uint32_t                  colorSlm    = 0;
-  uint32_t                  colorAod    = 1;
-  uint32_t                  colorLocal  = 2;
+  uint32_t colorSlm = 0;
+  uint32_t colorAod = 1;
+  uint32_t colorLocal = 2;
   [[maybe_unused]] uint32_t colorGlobal = 3;
-  uint32_t                  colorCz     = 4;
+  uint32_t colorCz = 4;
 
-  std::map<CoordIndex, HwQubit>                coordIdxToId;
+  std::map<CoordIndex, HwQubit> coordIdxToId;
   std::map<HwQubit, std::pair<qc::fp, qc::fp>> idToCoord;
-  std::map<HwQubit, uint32_t>                  axesIds;
-  std::map<HwQubit, uint32_t>                  marginIds;
-  uint32_t                                     axesIdCounter   = 0;
-  uint32_t                                     marginIdCounter = 0;
+  std::map<HwQubit, uint32_t> axesIds;
+  std::map<HwQubit, uint32_t> marginIds;
+  uint32_t axesIdCounter = 0;
+  uint32_t marginIdCounter = 0;
 
-  axesId   addAxis(HwQubit id);
-  void     removeAxis(HwQubit id) { axesIds.erase(id); }
+  axesId addAxis(HwQubit id);
+  void removeAxis(HwQubit id) { axesIds.erase(id); }
   marginId addMargin(HwQubit id);
-  void     removeMargin(HwQubit id) { marginIds.erase(id); }
+  void removeMargin(HwQubit id) { marginIds.erase(id); }
 
 public:
   AnimationAtoms(const std::map<HwQubit, HwQubit>& initHwPos,
-                 const NeutralAtomArchitecture&    arch);
+                 const NeutralAtomArchitecture& arch);
 
-  std::string        getInitString();
-  std::string        getEndString(qc::fp endTime);
+  std::string getInitString();
+  std::string getEndString(qc::fp endTime);
   static std::string createCsvLine(qc::fp startTime, HwQubit id, qc::fp x,
                                    qc::fp y, uint32_t size = 1,
                                    uint32_t color = 0, bool axes = false,
                                    axesId axId = 0, bool margin = false,
-                                   marginId marginId   = 0,
-                                   qc::fp   marginSize = 0);
-  std::string        createCsvOp(const std::unique_ptr<qc::Operation>& op,
-                                 qc::fp startTime, qc::fp endTime,
-                                 const NeutralAtomArchitecture& arch);
+                                   marginId marginId = 0,
+                                   qc::fp marginSize = 0);
+  std::string createCsvOp(const std::unique_ptr<qc::Operation>& op,
+                          qc::fp startTime, qc::fp endTime,
+                          const NeutralAtomArchitecture& arch);
 };
 
 } // namespace na
