@@ -6,10 +6,10 @@
 #pragma once
 
 #include "Definitions.hpp"
-#include "Permutation.hpp"
 #include "hybridmap/NeutralAtomDefinitions.hpp"
 #include "hybridmap/NeutralAtomUtils.hpp"
-#include "operations/Operation.hpp"
+#include "ir/Permutation.hpp"
+#include "ir/operations/Operation.hpp"
 
 #include <algorithm>
 #include <cstddef>
