@@ -6,7 +6,6 @@
 #include <cmath>
 #include <cstddef>
 #include <cstdint>
-#include <set>
 #include <utility>
 #include <vector>
 
@@ -18,14 +17,14 @@ struct NestedVar {
   explicit NestedVar(const LogicTerm& v) : var(v) {};
   NestedVar(const LogicTerm& v, std::vector<NestedVar> l)
       : var(v), list(std::move(l)) {}
-  LogicTerm              var = LogicTerm::noneTerm();
+  LogicTerm var = LogicTerm::noneTerm();
   std::vector<NestedVar> list;
 };
 
 struct WeightedVar {
   WeightedVar(const LogicTerm& v, const int w) : var(v), weight(w) {}
-  LogicTerm var    = LogicTerm::noneTerm();
-  int       weight = 0;
+  LogicTerm var = LogicTerm::noneTerm();
+  int weight = 0;
 };
 inline bool operator<(const WeightedVar& rhs, const WeightedVar& lhs) {
   return rhs.weight < lhs.weight;
@@ -38,8 +37,8 @@ enum class Type : uint8_t { Uninitialized, AuxVar, ProgramVar };
 struct SavedLit {
   SavedLit() : var(LogicTerm::noneTerm()) {}
   SavedLit(Type t, const LogicTerm& v) : type(t), var(v) {}
-  Type      type = Type::Uninitialized;
-  LogicTerm var  = LogicTerm::noneTerm();
+  Type type = Type::Uninitialized;
+  LogicTerm var = LogicTerm::noneTerm();
 };
 
 LogicTerm atMostOneCmdr(const std::vector<NestedVar>& subords,
@@ -55,12 +54,12 @@ LogicTerm naiveAtMostOne(const std::vector<LogicTerm>& clauseVars);
 LogicTerm naiveAtLeastOne(const std::vector<LogicTerm>& clauseVars);
 
 LogicTerm atMostOneBiMander(const std::vector<LogicTerm>& vars,
-                            LogicBlock*                   logic);
+                            LogicBlock* logic);
 
 std::vector<NestedVar> groupVars(const std::vector<LogicTerm>& vars,
-                                 std::size_t                   maxSize);
+                                 std::size_t maxSize);
 std::vector<NestedVar> groupVarsAux(const std::vector<NestedVar>& vars,
-                                    std::size_t                   maxSize);
+                                    std::size_t maxSize);
 
 std::vector<std::vector<LogicTerm>>
 groupVarsBimander(const std::vector<LogicTerm>& vars, std::size_t groupCount);
