@@ -16,9 +16,9 @@ public:
   explicit Model(Result res) : result(res) {}
   virtual ~Model() = default;
 
-  virtual int      getIntValue(const LogicTerm& a, LogicBlock* lb)       = 0;
-  virtual bool     getBoolValue(const LogicTerm& a, LogicBlock* lb)      = 0;
-  virtual double   getRealValue(const LogicTerm& a, LogicBlock* lb)      = 0;
+  virtual int getIntValue(const LogicTerm& a, LogicBlock* lb) = 0;
+  virtual bool getBoolValue(const LogicTerm& a, LogicBlock* lb) = 0;
+  virtual double getRealValue(const LogicTerm& a, LogicBlock* lb) = 0;
   virtual uint64_t getBitvectorValue(const LogicTerm& a, LogicBlock* lb) = 0;
 };
 } // namespace logicbase
