@@ -1,9 +1,3 @@
-//
-// This file is part of the MQT QMAP library released under the MIT license.
-// See README.md or go to https://github.com/cda-tum/mqt-qmap for more
-// information.
-//
-
 #pragma once
 
 #include <algorithm>
@@ -11,8 +5,10 @@
 #include <cstdint>
 #include <istream>
 #include <sstream>
+#include <stdexcept>
 #include <string>
 #include <unordered_map>
+
 namespace na {
 
 enum class NAMappingMethod : std::uint8_t {
@@ -38,9 +34,9 @@ getMethodOfString(const std::string& method) -> NAMappingMethod {
 }
 class Configuration {
 private:
-  std::size_t     patchRows = 1;
-  std::size_t     patchCols = 1;
-  NAMappingMethod method    = NAMappingMethod::MaximizeParallelismHeuristic;
+  std::size_t patchRows = 1;
+  std::size_t patchCols = 1;
+  NAMappingMethod method = NAMappingMethod::MaximizeParallelismHeuristic;
 
 public:
   Configuration() = default;
