@@ -9,10 +9,10 @@
 #include "../Configuration.hpp"
 #include "Architecture.hpp"
 #include "Definitions.hpp"
-#include "QuantumComputation.hpp"
+#include "ir/QuantumComputation.hpp"
+#include "ir/operations/Operation.hpp"
 #include "na/NAComputation.hpp"
 #include "na/NADefinitions.hpp"
-#include "operations/Operation.hpp"
 
 #include <cstddef>
 #include <cstdint>
@@ -30,15 +30,15 @@ namespace na {
 class NAMapper {
 public:
   struct Statistics {
-    std::size_t               numInitialGates    = 0;
-    std::size_t               numEntanglingGates = 0;
-    std::size_t               initialDepth       = 0;
-    std::size_t               numMappedGates     = 0;
-    std::size_t               numQubits          = 0;
-    std::size_t               maxSeqWidth        = 0;
-    qc::fp                    preprocessTime     = 0.0; // [ms]
-    qc::fp                    mappingTime        = 0.0; // [ms]
-    qc::fp                    postprocessTime    = 0.0; // [ms]
+    std::size_t numInitialGates = 0;
+    std::size_t numEntanglingGates = 0;
+    std::size_t initialDepth = 0;
+    std::size_t numMappedGates = 0;
+    std::size_t numQubits = 0;
+    std::size_t maxSeqWidth = 0;
+    qc::fp preprocessTime = 0.0;  // [ms]
+    qc::fp mappingTime = 0.0;     // [ms]
+    qc::fp postprocessTime = 0.0; // [ms]
     [[nodiscard]] static auto header() -> std::string {
       return "numInitialGates,numEntanglingGates,initialDepth,numMappedGates,"
              "numQubits,maxSeqWidth,preprocessTime,mappingTime,"
@@ -52,7 +52,7 @@ public:
          << postprocessTime << '\n';
       return ss.str();
     }
-    friend auto operator<<(std::ostream&     os,
+    friend auto operator<<(std::ostream& os,
                            const Statistics& s) -> std::ostream& {
       return os << s.toString();
     }
@@ -60,20 +60,20 @@ public:
 
 protected:
   qc::QuantumComputation initialQc;
-  NAComputation          mappedQc;
-  Architecture           initialArch;
-  Architecture           arch;
-  Configuration          config;
-  Statistics             stats{};
-  bool                   done = false;
+  NAComputation mappedQc;
+  Architecture initialArch;
+  Architecture arch;
+  Configuration config;
+  Statistics stats{};
+  bool done = false;
 
   class Atom {
   public:
     enum class PositionStatus : std::uint8_t { UNDEFINED, DEFINED };
-    PositionStatus         positionStatus  = PositionStatus::UNDEFINED;
+    PositionStatus positionStatus = PositionStatus::UNDEFINED;
     std::shared_ptr<Point> initialPosition = std::make_shared<Point>(0, 0);
     std::shared_ptr<Point> currentPosition = initialPosition;
-    std::vector<Zone>      zones;
+    std::vector<Zone> zones;
     explicit Atom(const std::vector<Zone>& z = {}) : zones(z) {};
   };
   auto preprocess() -> void { validateCircuit(); }
@@ -85,14 +85,14 @@ protected:
   auto makeLogicalArrays() -> void;
   auto calculateMovements() -> void;
   [[nodiscard]] auto
-       checkApplicability(const qc::Operation*     op,
-                          const std::vector<Atom>& placement) const -> bool;
+  checkApplicability(const qc::Operation* op,
+                     const std::vector<Atom>& placement) const -> bool;
   auto updatePlacement(const qc::Operation* op,
-                       std::vector<Atom>&   placement) const -> void;
+                       std::vector<Atom>& placement) const -> void;
   [[nodiscard]] static auto
-  getMisplacement(const std::vector<Atom>&      initial,
+  getMisplacement(const std::vector<Atom>& initial,
                   const std::vector<qc::Qubit>& target,
-                  const qc::Qubit&              q) -> std::int64_t;
+                  const qc::Qubit& q) -> std::int64_t;
   /**
    * @brief Move atoms from the entangling zone to the destination zone.
    * @param initialFreeSites The sites that are not yet occupied from the start
@@ -122,15 +122,15 @@ protected:
   auto pickUp(std::vector<bool>& initialFreeSites,
               std::vector<bool>& currentFreeSites, std::vector<Atom>& placement,
               std::unordered_set<qc::Qubit>& currentlyShuttling,
-              const std::vector<qc::Qubit>&  qubits) -> void;
+              const std::vector<qc::Qubit>& qubits) -> void;
 
 public:
-  explicit NAMapper(Architecture         architecture,
+  explicit NAMapper(Architecture architecture,
                     const Configuration& configuration)
       : initialArch(std::move(architecture)),
         arch(initialArch.withConfig(configuration)), config(configuration) {}
   virtual ~NAMapper() = default;
-  auto               map(const qc::QuantumComputation& qc) -> void;
+  auto map(const qc::QuantumComputation& qc) -> void;
   [[nodiscard]] auto getResult() const -> const NAComputation& {
     if (!done) {
       throw std::logic_error("No result available.");
