@@ -5,18 +5,26 @@
 
 #include "cliffordsynthesis/Tableau.hpp"
 
-#include "operations/OpType.hpp"
-#include "operations/Operation.hpp"
-#include "plog/Log.h"
-#include "utils.hpp"
+#include "ir/QuantumComputation.hpp"
+#include "ir/operations/CompoundOperation.hpp"
+#include "ir/operations/OpType.hpp"
+#include "ir/operations/Operation.hpp"
 
+#include <algorithm>
+#include <cassert>
+#include <cctype>
 #include <cstddef>
 #include <fstream>
 #include <istream>
+#include <optional>
 #include <ostream>
+#include <plog/Log.h>
 #include <regex>
+#include <set>
+#include <sstream>
 #include <stdexcept>
 #include <string>
+#include <utility>
 #include <vector>
 
 namespace cs {
@@ -42,19 +50,50 @@ void Tableau::import(const std::string& filename) {
   import(is);
 }
 
+void parseLine(const std::string& line, char separator,
+               const std::set<char>& escapeChars,
+               const std::set<char>& ignoredChars,
+               std::vector<std::string>& result) {
+  result.clear();
+  std::string word;
+  bool inEscape = false;
+  for (const char c : line) {
+    if (ignoredChars.find(c) != ignoredChars.end()) {
+      continue;
+    }
+    if (inEscape) {
+      if (escapeChars.find(c) != escapeChars.end()) {
+        inEscape = false;
+      } else {
+        word += c;
+      }
+    } else {
+      if (escapeChars.find(c) != escapeChars.end()) {
+        inEscape = true;
+      } else if (c == separator) {
+        result.push_back(word);
+        word = "";
+      } else {
+        word += c;
+      }
+    }
+  }
+  result.push_back(word);
+}
+
 void Tableau::import(std::istream& is) {
   tableau.clear();
 
-  std::string              line;
+  std::string line;
   std::vector<std::string> data{};
-  char                     delimiter = '|';
+  char delimiter = '|';
 
   while (std::getline(is, line)) {
     if (line.find('|', 0) == std::string::npos) {
       delimiter = ';';
     }
     tableau.emplace_back();
-    ::parseLine(line, delimiter, {'\"'}, {'\\', '\r', '\n', '\t'}, data);
+    parseLine(line, delimiter, {'\"'}, {'\\', '\r', '\n', '\t'}, data);
     for (const auto& datum : data) {
       if (datum.empty()) {
         continue;
@@ -154,7 +193,7 @@ void Tableau::applyGate(const qc::Operation* const gate) {
 }
 
 void Tableau::createDiagonalTableau(const std::size_t nQ,
-                                    const bool        includeDestabilizers) {
+                                    const bool includeDestabilizers) {
   nQubits = nQ;
   tableau.clear();
   if (includeDestabilizers) {
@@ -200,7 +239,7 @@ std::string Tableau::toString() const {
 
 void Tableau::fromString(const std::string& str) {
   std::stringstream ss(str);
-  std::string       line;
+  std::string line;
   std::getline(ss, line);
   if (line.empty()) {
     return;
@@ -291,7 +330,7 @@ void Tableau::applyCX(const std::size_t control, const std::size_t target) {
     const auto zb = tableau[i][target + nQubits];
     tableau[i][2 * nQubits] ^= (xa & zb) & ((xb ^ za) ^ 1);
     tableau[i][control + nQubits] = za ^ zb;
-    tableau[i][target]            = xb ^ xa;
+    tableau[i][target] = xb ^ xa;
   }
 }
 
@@ -391,7 +430,7 @@ Tableau::RowType Tableau::parseStabilizer(const std::string& stab) {
     if (stabCopy[stabCopy.size() - 1] == '\'') {
       stabCopy = stabCopy.substr(1, stabCopy.size() - 2);
     } else {
-      throw QMAPException("Unmatched \"'\" in stabilizer string");
+      throw std::invalid_argument("Unmatched \"'\" in stabilizer string");
     }
   }
   if (stabCopy[0] == '+' || stabCopy[0] == '-') {
@@ -405,9 +444,10 @@ Tableau::RowType Tableau::parseStabilizer(const std::string& stab) {
     } else if (c == 'X' || c == 'Y') {
       row.push_back(1);
     } else {
-      throw QMAPException("Invalid stabilizer " + stab +
-                          ". Stabilizers must be given as a list of stabilizer "
-                          "like [XYZI, ZIXZ]");
+      throw std::invalid_argument(
+          "Invalid stabilizer " + stab +
+          ". Stabilizers must be given as a list of stabilizer "
+          "like [XYZI, ZIXZ]");
     }
   }
   for (const auto c : stabCopy) {
@@ -416,9 +456,10 @@ Tableau::RowType Tableau::parseStabilizer(const std::string& stab) {
     } else if (c == 'Y' || c == 'Z') {
       row.push_back(1);
     } else {
-      throw QMAPException("Invalid stabilizer" + stab +
-                          ". Stabilizers must be given as a list of stabilizer "
-                          "like [XYZI, ZIXZ]");
+      throw std::invalid_argument(
+          "Invalid stabilizer" + stab +
+          ". Stabilizers must be given as a list of stabilizer "
+          "like [XYZI, ZIXZ]");
     }
   }
   if (stab[0U] == '-' || (stab[0U] == '\'' && stab[1U] == '-')) {
@@ -431,7 +472,7 @@ Tableau::RowType Tableau::parseStabilizer(const std::string& stab) {
 
 void Tableau::loadStabilizerDestabilizerString(const std::string& string) {
   std::stringstream ss(string);
-  std::string       line;
+  std::string line;
   std::getline(ss, line);
   if (line.empty()) {
     return;
@@ -445,25 +486,25 @@ void Tableau::loadStabilizerDestabilizerString(const std::string& string) {
     if (stabilizers[stabilizers.size() - 1] == ']') {
       stabilizers = stabilizers.substr(1, stabilizers.size() - 2);
     } else {
-      throw QMAPException("Unmatched \"[\" in stabilizer string");
+      throw std::invalid_argument("Unmatched \"[\" in stabilizer string");
     }
   }
 
   std::optional<std::size_t> stabLength;
-  const auto&                checkStabLength = [&](const RowType& row) {
+  const auto& checkStabLength = [&](const RowType& row) {
     if (!stabLength.has_value()) {
       stabLength = row.size();
     }
     if (stabLength.value() != row.size()) {
-      throw QMAPException("All Stabilizers muts have the same length");
+      throw std::invalid_argument("All Stabilizers muts have the same length");
     }
   };
 
-  const char  delimiter = ',';
+  const char delimiter = ',';
   std::string stab;
   for (std::size_t pos = stabilizers.find(delimiter); pos != std::string::npos;
-       pos             = stabilizers.find(delimiter)) {
-    stab            = stabilizers.substr(0, pos);
+       pos = stabilizers.find(delimiter)) {
+    stab = stabilizers.substr(0, pos);
     const auto& row = parseStabilizer(stab);
     checkStabLength(row);
     tableau.push_back(row);
