@@ -1,17 +1,14 @@
-//
-// This file is part of the MQT QMAP library released under the MIT license.
-// See README.md or go to https://github.com/cda-tum/mqt-qmap for more
-// information.
-//
+#include "na/Configuration.hpp"
 
-#include "Configuration.hpp"
-
+#include <exception>
 #include <fstream>
 #include <iostream>
 #include <nlohmann/json.hpp>
 #include <sstream>
+#include <stdexcept>
+#include <string>
 
-using json = nlohmann::json;
+using json = nlohmann::basic_json<>;
 
 namespace na {
 
@@ -31,7 +28,7 @@ Configuration::Configuration(std::istream& fs) {
     fs >> data;
   } catch (const std::exception& e) {
     std::stringstream ss;
-    ss << "Error parsing JSON: " << e.what() << std::endl;
+    ss << "Error parsing JSON: " << e.what() << '\n';
     throw std::runtime_error(ss.str());
   }
 
