@@ -1,17 +1,17 @@
-#include "Architecture.hpp"
-#include "Configuration.hpp"
 #include "Definitions.hpp"
-#include "NAMapper.hpp"
-#include "QuantumComputation.hpp"
 #include "datastructures/Layer.hpp"
+#include "ir/QuantumComputation.hpp"
+#include "ir/operations/CompoundOperation.hpp"
+#include "ir/operations/OpType.hpp"
+#include "ir/operations/StandardOperation.hpp"
+#include "na/Architecture.hpp"
+#include "na/Configuration.hpp"
 #include "na/NAComputation.hpp"
+#include "na/NAMapper.hpp"
 #include "na/NADefinitions.hpp"
 #include "na/operations/NAGlobalOperation.hpp"
 #include "na/operations/NALocalOperation.hpp"
 #include "na/operations/NAShuttlingOperation.hpp"
-#include "operations/CompoundOperation.hpp"
-#include "operations/OpType.hpp"
-#include "operations/StandardOperation.hpp"
 
 #include <algorithm>
 #include <cstddef>
@@ -27,10 +27,10 @@
 namespace na {
 
 auto retrieveQuantumComputation(const NAComputation& nac,
-                                const Architecture&  arch)
+                                const Architecture& arch)
     -> qc::QuantumComputation {
-  qc::QuantumComputation               qComp(nac.getInitialPositions().size());
-  std::vector<Point>                   positionOfQubits;
+  qc::QuantumComputation qComp(nac.getInitialPositions().size());
+  std::vector<Point> positionOfQubits;
   std::unordered_map<Point, qc::Qubit> positionToQubit;
   positionOfQubits.reserve(nac.getInitialPositions().size());
   qc::Qubit n = 0;
@@ -101,15 +101,15 @@ auto retrieveQuantumComputation(const NAComputation& nac,
 }
 
 auto checkEquivalence(const qc::QuantumComputation& circ,
-                      const NAComputation&          nac,
-                      const Architecture&           arch) -> bool {
-  auto            naQComp = retrieveQuantumComputation(nac, arch);
+                      const NAComputation& nac,
+                      const Architecture& arch) -> bool {
+  auto naQComp = retrieveQuantumComputation(nac, arch);
   const qc::Layer qLayer(circ);
-  int             line = 0;
+  int line = 0;
   for (const auto& op : naQComp) {
     ++line;
     const auto& executableSet = qLayer.getExecutableSet();
-    const auto& it            = std::find_if(
+    const auto& it = std::find_if(
         executableSet.begin(), executableSet.end(),
         [&op](const std::shared_ptr<qc::Layer::DAGVertex>& vertex) {
           return *vertex->getOperation() == *op;
@@ -235,7 +235,7 @@ TEST(NAMapper, Exceptions) {
           }
       ]
   })");
-  std::stringstream  gridSS;
+  std::stringstream gridSS;
   gridSS << "x,y\n";
   // entangling zone (4 x 36 = 144 sites)
   for (std::size_t y = 0; y <= 36; y += 12) {
@@ -389,7 +389,7 @@ TEST(NAMapper, QAOA10) {
           }
       ]
   })");
-  std::stringstream  gridSS;
+  std::stringstream gridSS;
   gridSS << "x,y\n";
   // entangling zone (4 x 36 = 144 sites)
   for (std::size_t y = 0; y <= 36; y += 12) {
@@ -670,8 +670,8 @@ rz(3.9927041) q[3];
 rz(3.9927041) q[4];
 rz(3.9927041) q[5];
 rz(3.9927041) q[7];)";
-  const auto&       circ = qc::QuantumComputation::fromQASM(qasm);
-  const auto&       arch = na::Architecture(archIS, gridSS);
+  const auto& circ = qc::QuantumComputation::fromQASM(qasm);
+  const auto& arch = na::Architecture(archIS, gridSS);
   // ---------------------------------------------------------------------
   na::NAMapper mapper(
       arch, na::Configuration(
@@ -803,7 +803,7 @@ TEST(NAMapper, QAOA16Narrow) {
       ]
   }
   )");
-  std::stringstream  gridSS;
+  std::stringstream gridSS;
   gridSS << "x,y\n";
   // entangling zone (4 x 36 = 144 sites)
   for (std::size_t y = 0; y <= 36; y += 12) {
@@ -923,8 +923,8 @@ ry(0.3223291) q;
 cp(pi) q[9],q[11];
 ry(-2.2154814) q;
 ry(2.2154814) q;)";
-  const auto&       circ = qc::QuantumComputation::fromQASM(qasm);
-  const auto&       arch = na::Architecture(archIS, gridSS);
+  const auto& circ = qc::QuantumComputation::fromQASM(qasm);
+  const auto& arch = na::Architecture(archIS, gridSS);
   // ---------------------------------------------------------------------
   na::NAMapper mapper(
       arch, na::Configuration(
@@ -1039,7 +1039,7 @@ TEST(NAMapper, QAOA16NarrowEntangling) {
       ]
   }
   )");
-  std::stringstream  gridSS;
+  std::stringstream gridSS;
   gridSS << "x,y\n";
   // entangling zone (4 x 36 = 144 sites)
   for (std::size_t y = 0; y <= 36; y += 12) {
@@ -1159,8 +1159,8 @@ ry(0.3223291) q;
 cp(pi) q[9],q[11];
 ry(-2.2154814) q;
 ry(2.2154814) q;)";
-  const auto&       circ = qc::QuantumComputation::fromQASM(qasm);
-  const auto&       arch = na::Architecture(archIS, gridSS);
+  const auto& circ = qc::QuantumComputation::fromQASM(qasm);
+  const auto& arch = na::Architecture(archIS, gridSS);
   // ---------------------------------------------------------------------
   na::NAMapper mapper(
       arch, na::Configuration(
